requirements-dev.txt
-r requirements.txt

# Additional development tools
pytest-watch>=4.2.0
pytest-xdist>=3.0.0
pytest-sugar>=0.9.7
pre-commit>=3.0.0
bandit>=1.7.0
safety>=2.3.0

# Documentation
Sphinx>=7.0.0
sphinx-rtd-theme>=1.0.0
sphinx-autodoc-typehints>=1.22.0

# Debug tools
ipython>=8.0.0
ipdb>=0.13.0
